#629: Leverage BigDecimal classes for decimals calculations

+++ src/main/java/org/benetech/servicenet/matching/counter/EmailSimilarityCounter.java
@@ -1,5 +1,6 @@
 package org.benetech.servicenet.matching.counter;
 
+import java.math.BigDecimal;
 import org.apache.commons.lang3.StringUtils;
 import org.benetech.servicenet.matching.model.MatchingContext;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,13 +16,13 @@ public class EmailSimilarityCounter extends AbstractSimilarityCounter<String> {
     private static final int LOCAL_PART = 0;
 
     @Value("${similarity-ratio.weight.email.same-domain}")
-    private float domainWeight;
+    private BigDecimal domainWeight;
 
     @Value("${similarity-ratio.weight.email.same-normalized-local-parts}")
-    private float normalizedLocalPartWeight;
+    private BigDecimal normalizedLocalPartWeight;
 
     @Override
-    public float countSimilarityRatio(String email1, String email2, MatchingContext context) {
+    public BigDecimal countSimilarityRatio(String email1, String email2, MatchingContext context) {
         if (StringUtils.isBlank(email1) || StringUtils.isBlank(email2)) {
             return NO_MATCH_RATIO;
         }
